Add ability to download plugins locally, in order to speed up jenkins build time

+++ jenkins/plugins.sh
@@ -10,6 +10,7 @@
 
 set -e
 
+REF_LOCAL=/usr/share/jenkins/ref/plugins-local
 REF=/usr/share/jenkins/ref/plugins
 mkdir -p $REF
 
@@ -18,11 +19,17 @@ while read spec || [ -n "$spec" ]; do
     [[ ${plugin[0]} =~ ^# ]] && continue
     [[ ${plugin[0]} =~ ^\s*$ ]] && continue
     [[ -z ${plugin[1]} ]] && plugin[1]="latest"
-    echo "Downloading ${plugin[0]}:${plugin[1]}"
 
     if [ -z "$JENKINS_UC_DOWNLOAD" ]; then
       JENKINS_UC_DOWNLOAD=$JENKINS_UC/download
     fi
-    curl -sSL -f ${JENKINS_UC_DOWNLOAD}/plugins/${plugin[0]}/${plugin[1]}/${plugin[0]}.hpi -o $REF/${plugin[0]}.jpi
-    unzip -qqt $REF/${plugin[0]}.jpi
+    if [ -e "${REF_LOCAL}/${plugin[0]}-${plugin[1]}.jpi" ]; then
+        echo "Plugin ${plugin[0]}-${plugin[1]}.jpi exists locally, it will not be downloaded from ${JENKINS_UC_DOWNLOAD}"
+        cp $REF_LOCAL/${plugin[0]}-${plugin[1]}.jpi $REF/${plugin[0]}.jpi
+    else
+        echo "Plugin does not exist locally, it will be downloaded from ${JENKINS_UC_DOWNLOAD}"
+        echo "Downloading ${plugin[0]}:${plugin[1]}"
+        curl -sSL -f ${JENKINS_UC_DOWNLOAD}/plugins/${plugin[0]}/${plugin[1]}/${plugin[0]}.hpi -o $REF/${plugin[0]}.jpi
+        unzip -qqt $REF/${plugin[0]}.jpi
+    fi
 done  < $1
